+++ ui/mods/com.pa.quitch.qquellerai/new_game.js
@@ -147,13 +147,14 @@ function qQuellerAI() {
       go_for_the_kill: true,
       priority_scout_metal_spots: true,
       enable_commander_danger_responses: true,
-      neural_data_mod: 1.25,
+      neural_data_mod: 1.0,
       adv_eco_mod: 1.0,
       adv_eco_mod_alone: 0,
       personality_tags:
         [
           "uber",
-          "aggressive"
+          "aggressive",
+          "neural"
         ],
       min_basic_fabbers: 3,
       min_advanced_fabbers: 1,
@@ -242,27 +243,6 @@ function qQuellerAI() {
       min_basic_fabbers: 1,
       min_advanced_fabbers: 1
     },
-    'Q-Uber Neural': {
-      display_name: '!LOC:Q-Uber Neural',
-      metal_drain_check: 0.54,
-      energy_drain_check: 0.65,
-      metal_demand_check: 0.85,
-      energy_demand_check: 0.8,
-      micro_type: 2,
-      go_for_the_kill: true,
-      priority_scout_metal_spots: true,
-      enable_commander_danger_responses: true,
-      neural_data_mod: 1.0,
-      adv_eco_mod: 1.0,
-      adv_eco_mod_alone: 0,
-      personality_tags:
-        [
-          "uber",
-          "neural"
-        ],
-      min_basic_fabbers: 3,
-      min_advanced_fabbers: 1
-    },
     'Q-Uber Rush': {
       display_name: '!LOC:Q-Uber Rush',
       metal_drain_check: 0.54,
@@ -279,7 +259,7 @@ function qQuellerAI() {
       personality_tags:
         [
           "uber",
-          "rush"
+          "aggressive"
         ],
       min_basic_fabbers: 2,
       min_advanced_fabbers: 1
@@ -320,7 +300,7 @@ function qQuellerAI() {
   model.aiPersonalities.valueHasMutated();
 
   _.defer(function () {
-    model.localChatMessage(loc("!LOC:Queller AI"), loc("!LOC:by Quitch. Version 4.46.3. My difficulties are prefixed with a Q. Don't make AI teams larger than human teams, increase the difficulty or the econ rate instead. I support the Legion Expansion."));
+    model.localChatMessage(loc("!LOC:Queller AI"), loc("!LOC:by Quitch. Version 4.47.0. My difficulties are prefixed with a Q. Don't make AI teams larger than human teams, increase the difficulty or the econ rate instead. I support the Legion Expansion."));
   });
 
 }